tox.ini file
tox is a nice python utility to run tests, it let you define isolated
environnement which run specific list of commands.

This patches defines the pep8 and pyflakes env which are both linting
tools. To run then run:

 tox -v -epep8
 tox -v -epyflakes

Or simply:

 tox

You will need the python-tox packages to be installed or get it from
pip.

Change-Id: Ief41419a7aa6fd4d69808b097f1e6b3a0f1d4b37