console script now uses the new 'cli' entry point
sartoris 'main' entry points require an argument which is not possible
using setuptools.  I am defining a new entry point named 'cli' which
takes no argument and in turn call 'main' with the system arguments.

The 'main' entry point is left unchanged and still require argument to
be passed to it.

Test plan:

	$ sudo python setup.py develop
	$ sartoris
	usage: sartoris [-q --quiet] [-s --silent] [-v --verbose] [method]
	sartoris: error: too few arguments
	$

Change-Id: Ia805543d6e3718f253caa0f1c1f815bb0a085724